Type cast compile warning fix.

+++ net_interface.cpp
@@ -13,5 +13,5 @@ u_long iface_RTT(const struct net_interface& local_iface,
                  const struct net_interface& remote_iface)
 {
     double rtt = 2*((local_iface.RTT / 2.0) + (remote_iface.RTT / 2.0));
-    return rtt;
+    return (u_long)rtt;
 }
